fake 데이터 삭제 및 실제 api 데이터 시작

+++ lib/repository/fishing_api.dart
@@ -223,8 +223,8 @@ class Items {
     _fshlcType = json["fshlcType"];
     _rdnmadr = json["rdnmadr"];
     _lnmadr = json["lnmadr"];
-    _latitude = json["latitude"];
-    _longitude = json["longitude"];
+    _latitude = json["latitude"].toDouble();
+    _longitude = json["longitude"].toDouble();
     _fshlcPhoneNumber = json["fshlcPhoneNumber"];
     _waterAr = json["waterAr"];
     _kdfsh = json["kdfsh"];
